requirements.txt
asgiref
channels
django
example
pytz
six
sqlparse
emoji
pillow
requests
django-recaptcha